refactor: 패키징 및 컴퓨터 랜덤숫자 중복검사 추가
