+++ src/main/java/baseball/domain/Baseball.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public class Baseball {
     private String computer;
