+++ src/main/java/baseball/domain/BaseballEnum.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public enum BaseballEnum {
     STRIKE, BALL
